refactor: create an isolated test env for each test instead of test suite
The reason behind these change is that test were failing on a faster machine
(probably due to the implementation of KafkaJS and Kafka itself). The clean up
of topics were failing and consequently also the tests. An easier solution
is not to clean up the environment but to have a true isolated environment
for each test.